pyproject.toml
[project]
name = "PyArch"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "loguru (>=0.7.3,<0.8.0) ; python_version >= '3.13' and python_version < '4.0'",
    "httpx (==0.27.0)",
]

[tool.poetry]
packages = [{include = "src", from = "."}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"
pyright = "^1.1.397"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"
autoflake = "^2.3.1"
pytest = "8.2.2"
pytest-asyncio = "0.23.7"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.pyright]
exclude = [
	"**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs",
]
reportMissingImports = true
reportTypeshedErrors = false
reportMissingTypeStubs = false
strictParameterNoneValue = false
reportOptionalMemberAccess = "none"
reportUndefinedVariable = true
typeCheckingMode = "basic"
# typeCheckingMode = "strict"
